The 'database' directory has been renamed to 'data'
