+++ debug.lua
@@ -15,10 +15,10 @@ require "src.lua.database"
 require "src.lua.console"
 require "src.lua.init"
 require "config"
-require "database.pokemons"
-require "database.fakemons"
-require "database.items"
-require "database.misc"
+require "data.pokemons"
+require "data.fakemons"
+require "data.items"
+require "data.misc"
 require "src.lua.admin"
 require "src.lua.room"
 require "src.lua.translate"
